% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{load_TREK}
\alias{load_TREK}
\title{The length of a string}
\usage{
load_TREK()
}
\value{
Loads the TREK barcodes into memory.
}
\description{
A function to load TREK barcodes into memory.
}
\examples{
message('correct usage') #load_TREK()
}
